[Tapir] Code refactoring to support the option of lowering Tapir to the CilkR runtime system.
