+++ include/llvm/Transforms/Tapir/TapirTypes.h
@@ -22,7 +22,8 @@ enum class TapirTargetType {
   None = 0,
   Serial = 1,
   Cilk = 2,
-  OpenMP = 3
+  OpenMP = 3,
+  CilkR = 4,
 };
 } // end namespace tapir
 } // end namespace llvm
