+++ lib/IR/Type.cpp
@@ -353,7 +353,7 @@ StructType *StructType::get(LLVMContext &Context, ArrayRef<Type*> ETypes,
   return ST;
 }
 
-StructType *StructType::getOrCreate(LLVMContext &Context, StringRef Name) {
+StructType *StructType::lookupOrCreate(LLVMContext &Context, StringRef Name) {
   StructType *Ty = Context.pImpl->NamedStructTypes.lookup(Name);
   if (!Ty)
     Ty = StructType::create(Context, Name);
