+++ lib/Transforms/Tapir/LoopSpawning.cpp
@@ -59,14 +59,11 @@ using std::make_pair;
 using namespace llvm;
 using namespace llvm::tapir;
 
-#define LS_NAME "loop-spawning"
 #define DEBUG_TYPE LS_NAME
 
 STATISTIC(LoopsAnalyzed, "Number of Tapir loops analyzed");
 STATISTIC(LoopsConvertedToDAC,
           "Number of Tapir loops converted to divide-and-conquer iteration spawning");
-STATISTIC(LoopsConvertedToCilkABI,
-          "Number of Tapir loops converted to use the Cilk ABI for loops");
 
 static cl::opt<TapirTargetType> ClTapirTarget(
     "ls-tapir-target", cl::desc("Target runtime for Tapir"),
@@ -135,68 +132,6 @@ static void emitMissedWarning(Function *F, Loop *L,
   }
 }
 
-/// LoopOutline serves as a base class for different variants of LoopSpawning.
-/// LoopOutline implements common parts of LoopSpawning transformations, namely,
-/// lifting a Tapir loop into a separate helper function.
-class LoopOutline {
-public:
-  LoopOutline(Loop *OrigLoop, ScalarEvolution &SE,
-              LoopInfo *LI, DominatorTree *DT,
-              AssumptionCache *AC,
-              OptimizationRemarkEmitter &ORE)
-      : OrigLoop(OrigLoop), SE(SE), LI(LI), DT(DT), AC(AC), ORE(ORE),
-        ExitBlock(nullptr)
-  {
-    // Use the loop latch to determine the canonical exit block for this loop.
-    TerminatorInst *TI = OrigLoop->getLoopLatch()->getTerminator();
-    if (2 != TI->getNumSuccessors())
-      return;
-    ExitBlock = TI->getSuccessor(0);
-    if (ExitBlock == OrigLoop->getHeader())
-      ExitBlock = TI->getSuccessor(1);
-  }
-
-  virtual bool processLoop() = 0;
-
-  virtual ~LoopOutline() {}
-
-protected:
-  PHINode* canonicalizeIVs(Type *Ty);
-  Value* canonicalizeLoopLatch(PHINode *IV, Value *Limit);
-  void unlinkLoop();
-
-  /// The original loop.
-  Loop *OrigLoop;
-
-  /// A wrapper around ScalarEvolution used to add runtime SCEV checks. Applies
-  /// dynamic knowledge to simplify SCEV expressions and converts them to a
-  /// more usable form.
-  // PredicatedScalarEvolution &PSE;
-  ScalarEvolution &SE;
-  /// Loop info.
-  LoopInfo *LI;
-  /// Dominator tree.
-  DominatorTree *DT;
-  /// Assumption cache.
-  AssumptionCache *AC;
-  /// Interface to emit optimization remarks.
-  OptimizationRemarkEmitter &ORE;
-
-  /// The exit block of this loop.  We compute our own exit block, based on the
-  /// latch, and handle other exit blocks (i.e., for exception handling) in a
-  /// special manner.
-  BasicBlock *ExitBlock;
-
-// private:
-//   /// Report an analysis message to assist the user in diagnosing loops that are
-//   /// not transformed.  These are handled as LoopAccessReport rather than
-//   /// VectorizationReport because the << operator of LoopSpawningReport returns
-//   /// LoopAccessReport.
-//   void emitAnalysis(const LoopAccessReport &Message) const {
-//     emitAnalysisDiag(OrigLoop, *ORE, Message);
-//   }
-};
-
 /// DACLoopSpawning implements the transformation to spawn the iterations of a
 /// Tapir loop in a recursive divide-and-conquer fashion.
 class DACLoopSpawning : public LoopOutline {
@@ -248,34 +183,6 @@ protected:
 //   }
 };
 
-/// CilkABILoopSpawning uses the Cilk Plus ABI to handle Tapir loops.
-class CilkABILoopSpawning : public LoopOutline {
-public:
-  CilkABILoopSpawning(Loop *OrigLoop, ScalarEvolution &SE,
-                      LoopInfo *LI, DominatorTree *DT,
-                      AssumptionCache *AC,
-                      OptimizationRemarkEmitter &ORE)
-      : LoopOutline(OrigLoop, SE, LI, DT, AC, ORE)
-  {}
-
-  bool processLoop();
-
-  virtual ~CilkABILoopSpawning() {}
-
-protected:
-  // PHINode* canonicalizeIVs(Type *Ty);
-  Value* canonicalizeLoopLatch(PHINode *IV, Value *Limit);
-
-// private:
-//   /// Report an analysis message to assist the user in diagnosing loops that are
-//   /// not transformed.  These are handled as LoopAccessReport rather than
-//   /// VectorizationReport because the << operator of LoopSpawningReport returns
-//   /// LoopAccessReport.
-//   void emitAnalysis(const LoopAccessReport &Message) const {
-//     emitAnalysisDiag(OrigLoop, *ORE, Message);
-//   }
-};
-
 struct LoopSpawningImpl {
   // LoopSpawningImpl(Function &F, LoopInfo &LI, ScalarEvolution &SE,
   //                  DominatorTree &DT,
@@ -293,7 +200,6 @@ struct LoopSpawningImpl {
   //     : F(F), GetLI(GetLI), LI(nullptr), GetSE(GetSE), GetDT(GetDT),
   //       ORE(ORE)
   // {}
-  TapirTarget* tapirTarget;
   LoopSpawningImpl(Function &F,
                    LoopInfo &LI,
                    ScalarEvolution &SE,
@@ -321,6 +227,8 @@ private:
   // AliasAnalysis *AA;
   AssumptionCache &AC;
   OptimizationRemarkEmitter &ORE;
+
+  TapirTarget* tapirTarget;
 };
 } // end anonymous namespace
 
@@ -1382,496 +1290,6 @@ bool DACLoopSpawning::processLoop() {
   }
 
   ++LoopsConvertedToDAC;
-
-  unlinkLoop();
-
-  return Helper;
-}
-
-/// \brief Replace the latch of the loop to check that IV is always less than or
-/// equal to the limit.
-///
-/// This method assumes that the loop has a single loop latch.
-Value* CilkABILoopSpawning::canonicalizeLoopLatch(PHINode *IV, Value *Limit) {
-  Loop *L = OrigLoop;
-
-  Value *NewCondition;
-  BasicBlock *Header = L->getHeader();
-  BasicBlock *Latch = L->getLoopLatch();
-  assert(Latch && "No single loop latch found for loop.");
-
-  IRBuilder<> Builder(&*Latch->getFirstInsertionPt());
-
-  // This process assumes that IV's increment is in Latch.
-
-  // Create comparison between IV and Limit at top of Latch.
-  NewCondition =
-    Builder.CreateICmpULT(Builder.CreateAdd(IV,
-                                            ConstantInt::get(IV->getType(), 1)),
-                          Limit);
-
-  // Replace the conditional branch at the end of Latch.
-  BranchInst *LatchBr = dyn_cast_or_null<BranchInst>(Latch->getTerminator());
-  assert(LatchBr && LatchBr->isConditional() &&
-         "Latch does not terminate with a conditional branch.");
-  Builder.SetInsertPoint(Latch->getTerminator());
-  Builder.CreateCondBr(NewCondition, Header, ExitBlock);
-
-  // Erase the old conditional branch.
-  LatchBr->eraseFromParent();
-
-  return NewCondition;
-}
-
-/// Top-level call to convert a Tapir loop to be processed using an appropriate
-/// Cilk ABI call.
-bool CilkABILoopSpawning::processLoop() {
-  Loop *L = OrigLoop;
-
-  BasicBlock *Header = L->getHeader();
-  BasicBlock *Preheader = L->getLoopPreheader();
-  BasicBlock *Latch = L->getLoopLatch();
-
-  using namespace ore;
-
-  // Check the exit blocks of the loop.
-  if (!ExitBlock) {
-    DEBUG(dbgs() << "LS loop does not contain valid exit block after latch.\n");
-    ORE.emit(OptimizationRemarkAnalysis(LS_NAME, "InvalidLatchExit",
-                                        L->getStartLoc(),
-                                        Header)
-             << "invalid latch exit");
-    return false;
-  }
-
-  SmallVector<BasicBlock *, 4> ExitBlocks;
-  L->getExitBlocks(ExitBlocks);
-  for (const BasicBlock *Exit : ExitBlocks) {
-    if (Exit == ExitBlock) continue;
-    if (!isa<UnreachableInst>(Exit->getTerminator())) {
-      DEBUG(dbgs() << "LS loop contains a bad exit block " << *Exit);
-      ORE.emit(OptimizationRemarkAnalysis(LS_NAME, "BadExit",
-                                          L->getStartLoc(),
-                                          Header)
-               << "bad exit block found");
-      return false;
-    }
-  }
-
-  Function *F = Header->getParent();
-  Module* M = F->getParent();
-
-  DEBUG(dbgs() << "LS loop header:" << *Header);
-  DEBUG(dbgs() << "LS loop latch:" << *Latch);
-
-  // DEBUG(dbgs() << "LS SE backedge taken count: " << *(SE.getBackedgeTakenCount(L)) << "\n");
-  // DEBUG(dbgs() << "LS SE max backedge taken count: " << *(SE.getMaxBackedgeTakenCount(L)) << "\n");
-  DEBUG(dbgs() << "LS SE exit count: " << *(SE.getExitCount(L, Latch)) << "\n");
-
-  /// Get loop limit.
-  const SCEV *BETC = SE.getExitCount(L, Latch);
-  const SCEV *Limit = SE.getAddExpr(BETC, SE.getOne(BETC->getType()));
-  DEBUG(dbgs() << "LS Loop limit: " << *Limit << "\n");
-  // PredicatedScalarEvolution PSE(SE, *L);
-  // const SCEV *PLimit = PSE.getExitCount(L, Latch);
-  // DEBUG(dbgs() << "LS predicated loop limit: " << *PLimit << "\n");
-  // emitAnalysis(LoopSpawningReport()
-  //              << "computed loop limit " << *Limit << "\n");
-  if (SE.getCouldNotCompute() == Limit) {
-    DEBUG(dbgs() << "SE could not compute loop limit.\n");
-    ORE.emit(OptimizationRemarkAnalysis(LS_NAME, "UnknownLoopLimit",
-                                        L->getStartLoc(),
-                                        Header)
-             << "could not compute limit");
-    return false;
-  }
-  // ORE.emit(OptimizationRemarkAnalysis(LS_NAME, "LoopLimit", L->getStartLoc(),
-  //                                     Header)
-  //          << "loop limit: " << NV("Limit", Limit));
-  /// Clean up the loop's induction variables.
-  PHINode *CanonicalIV = canonicalizeIVs(Limit->getType());
-  if (!CanonicalIV) {
-    DEBUG(dbgs() << "Could not get canonical IV.\n");
-    // emitAnalysis(LoopSpawningReport()
-    //              << "Could not get a canonical IV.\n");
-    ORE.emit(OptimizationRemarkAnalysis(LS_NAME, "NoCanonicalIV",
-                                        L->getStartLoc(),
-                                        Header)
-             << "could not find or create canonical IV");
-    return false;
-  }
-  const SCEVAddRecExpr *CanonicalSCEV =
-    cast<const SCEVAddRecExpr>(SE.getSCEV(CanonicalIV));
-
-  // Remove all IV's other can CanonicalIV.
-  // First, check that we can do this.
-  bool CanRemoveIVs = true;
-  for (BasicBlock::iterator II = Header->begin(); isa<PHINode>(II); ++II) {
-    PHINode *PN = cast<PHINode>(II);
-    if (CanonicalIV == PN) continue;
-    // dbgs() << "IV " << *PN;
-    const SCEV *S = SE.getSCEV(PN);
-    // dbgs() << " SCEV " << *S << "\n";
-    if (SE.getCouldNotCompute() == S) {
-      // emitAnalysis(LoopSpawningReport(PN)
-      //              << "Could not compute the scalar evolution.\n");
-      ORE.emit(OptimizationRemarkAnalysis(LS_NAME, "NoSCEV", PN)
-               << "could not compute scalar evolution of "
-               << NV("PHINode", PN));
-      CanRemoveIVs = false;
-    }
-  }
-
-  if (!CanRemoveIVs) {
-    DEBUG(dbgs() << "Could not compute scalar evolutions for all IV's.\n");
-    return false;
-  }
-
-  ////////////////////////////////////////////////////////////////////////
-  // We now have everything we need to extract the loop.  It's time to
-  // do some surgery.
-
-  SCEVExpander Exp(SE, M->getDataLayout(), "ls");
-
-  // Remove the IV's (other than CanonicalIV) and replace them with
-  // their stronger forms.
-  //
-  // TODO?: We can probably adapt this process such that we don't require all
-  // IV's to be canonical.
-  {
-    SmallVector<PHINode*, 8> IVsToRemove;
-    for (BasicBlock::iterator II = Header->begin(); isa<PHINode>(II); ++II) {
-      PHINode *PN = cast<PHINode>(II);
-      if (PN == CanonicalIV) continue;
-      const SCEV *S = SE.getSCEV(PN);
-      Value *NewIV = Exp.expandCodeFor(S, S->getType(), CanonicalIV);
-      PN->replaceAllUsesWith(NewIV);
-      IVsToRemove.push_back(PN);
-    }
-    for (PHINode *PN : IVsToRemove)
-      PN->eraseFromParent();
-  }
-
-  // All remaining IV's should be canonical.  Collect them.
-  //
-  // TODO?: We can probably adapt this process such that we don't require all
-  // IV's to be canonical.
-  SmallVector<PHINode*, 8> IVs;
-  bool AllCanonical = true;
-  for (BasicBlock::iterator II = Header->begin(); isa<PHINode>(II); ++II) {
-    PHINode *PN = cast<PHINode>(II);
-    DEBUG({
-        const SCEVAddRecExpr *PNSCEV =
-          dyn_cast<const SCEVAddRecExpr>(SE.getSCEV(PN));
-        assert(PNSCEV && "PHINode did not have corresponding SCEVAddRecExpr");
-        assert(PNSCEV->getStart()->isZero() &&
-               "PHINode SCEV does not start at 0");
-        dbgs() << "LS step recurrence for SCEV " << *PNSCEV << " is "
-               << *(PNSCEV->getStepRecurrence(SE)) << "\n";
-        assert(PNSCEV->getStepRecurrence(SE)->isOne() &&
-               "PHINode SCEV step is not 1");
-      });
-    if (ConstantInt *C =
-        dyn_cast<ConstantInt>(PN->getIncomingValueForBlock(Preheader))) {
-      if (C->isZero())
-        IVs.push_back(PN);
-    } else {
-      AllCanonical = false;
-      DEBUG(dbgs() << "Remaining non-canonical PHI Node found: " << *PN << "\n");
-      // emitAnalysis(LoopSpawningReport(PN)
-      //              << "Found a remaining non-canonical IV.\n");
-      ORE.emit(OptimizationRemarkAnalysis(DEBUG_TYPE, "NonCanonicalIV", PN)
-               << "found a remaining noncanonical IV");
-    }
-  }
-  if (!AllCanonical)
-    return false;
-
-  // Insert the computation for the loop limit into the Preheader.
-  Value *LimitVar = Exp.expandCodeFor(Limit, Limit->getType(),
-                                      Preheader->getTerminator());
-  DEBUG(dbgs() << "LimitVar: " << *LimitVar << "\n");
-
-  // Canonicalize the loop latch.
-  Value *NewCond = canonicalizeLoopLatch(CanonicalIV, LimitVar);
-
-  /// Clone the loop into a new function.
-
-  // Get the inputs and outputs for the Loop blocks.
-  SetVector<Value*> Inputs, Outputs;
-  SetVector<Value*> BodyInputs, BodyOutputs;
-  ValueToValueMapTy VMap, InputMap;
-  std::vector<BasicBlock *> LoopBlocks;
-  AllocaInst* closure;
-  // Add start iteration, end iteration, and grainsize to inputs.
-  {
-    LoopBlocks = L->getBlocks();
-    // // Add exit blocks terminated by unreachable.  There should not be any other
-    // // exit blocks in the loop.
-    // SmallSet<BasicBlock *, 4> UnreachableExits;
-    // for (BasicBlock *Exit : ExitBlocks) {
-    //   if (Exit == ExitBlock) continue;
-    //   assert(isa<UnreachableInst>(Exit->getTerminator()) &&
-    //          "Found problematic exit block.");
-    //   UnreachableExits.insert(Exit);
-    // }
-
-    // // Add unreachable and exception-handling exits to the set of loop blocks to
-    // // clone.
-    // for (BasicBlock *BB : UnreachableExits)
-    //   LoopBlocks.push_back(BB);
-    // for (BasicBlock *BB : EHExits)
-    //   LoopBlocks.push_back(BB);
-
-    // DEBUG({
-    //     dbgs() << "LoopBlocks: ";
-    //     for (BasicBlock *LB : LoopBlocks)
-    //       dbgs() << LB->getName() << "("
-    //              << *(LB->getTerminator()) << "), ";
-    //     dbgs() << "\n";
-    //   });
-
-    // Get the inputs and outputs for the loop body.
-    {
-      // CodeExtractor Ext(LoopBlocks, DT);
-      // Ext.findInputsOutputs(BodyInputs, BodyOutputs);
-      SmallPtrSet<BasicBlock *, 32> Blocks;
-      for (BasicBlock *BB : LoopBlocks)
-        Blocks.insert(BB);
-      findInputsOutputs(Blocks, BodyInputs, BodyOutputs);
-    }
-
-    // Add argument for start of CanonicalIV.
-    DEBUG({
-        Value *CanonicalIVInput =
-          CanonicalIV->getIncomingValueForBlock(Preheader);
-        // CanonicalIVInput should be the constant 0.
-        assert(isa<Constant>(CanonicalIVInput) &&
-               "Input to canonical IV from preheader is not constant.");
-      });
-    Argument *StartArg = new Argument(CanonicalIV->getType(),
-                                      CanonicalIV->getName()+".start");
-    Inputs.insert(StartArg);
-    InputMap[CanonicalIV] = StartArg;
-
-    // Add argument for end.
-    Value* ea;
-    if (isa<Constant>(LimitVar)) {
-      Argument *EndArg = new Argument(LimitVar->getType(), "end");
-      Inputs.insert(EndArg);
-      ea = InputMap[LimitVar] = EndArg;
-    } else {
-      Inputs.insert(LimitVar);
-      ea = InputMap[LimitVar] = LimitVar;
-    }
-
-    // Put all of the inputs together, and clear redundant inputs from
-    // the set for the loop body.
-    SmallVector<Value*, 8> BodyInputsToRemove;
-    SmallVector<Value*, 8> StructInputs;
-    SmallVector<Type*, 8> StructIT;
-    for (Value *V : BodyInputs) {
-      if (!Inputs.count(V)) {
-        StructInputs.push_back(V);
-        StructIT.push_back(V->getType());
-      }
-      else
-        BodyInputsToRemove.push_back(V);
-    }
-    StructType* ST = StructType::create(StructIT);
-    IRBuilder<> B(L->getLoopPreheader()->getTerminator());
-    IRBuilder<> B2(L->getHeader()->getFirstNonPHIOrDbgOrLifetime());
-    closure = B.CreateAlloca(ST);
-    for(unsigned i=0; i<StructInputs.size(); i++) {
-      B.CreateStore(StructInputs[i], B.CreateConstGEP2_32(ST, closure, 0, i));
-      auto l2 = B2.CreateLoad(B2.CreateConstGEP2_32(ST, closure, 0, i));
-      auto UI = StructInputs[i]->use_begin(), E = StructInputs[i]->use_end();
-      for (; UI != E;) {
-        Use &U = *UI;
-        ++UI;
-        auto *Usr = dyn_cast<Instruction>(U.getUser());
-        if (Usr && !L->contains(Usr->getParent()))
-          continue;
-        U.set(l2);
-      }
-    }
-    Inputs.insert(closure);
-    //llvm::errs() << "<B>\n";
-    //for(auto& a : Inputs) a->dump();
-    //llvm::errs() << "</B>\n";
-    //StartArg->dump();
-    //ea->dump();
-    Inputs.remove(StartArg);
-    Inputs.insert(StartArg);
-    Inputs.remove(ea);
-    Inputs.insert(ea);
-    //llvm::errs() << "<A>\n";
-    //for(auto& a : Inputs) a->dump();
-    //llvm::errs() << "</A>\n";
-    for (Value *V : BodyInputsToRemove)
-      BodyInputs.remove(V);
-    assert(0 == BodyOutputs.size() &&
-           "All results from parallel loop should be passed by memory already.");
-  }
-  DEBUG({
-      for (Value *V : Inputs)
-        dbgs() << "EL input: " << *V << "\n";
-      for (Value *V : Outputs)
-        dbgs() << "EL output: " << *V << "\n";
-    });
-
-
-  Function *Helper;
-  {
-    SmallVector<ReturnInst *, 4> Returns;  // Ignore returns cloned.
-
-    // LowerDbgDeclare(*(Header->getParent()));
-
-    Helper = CreateHelper(Inputs, Outputs, L->getBlocks(),
-                          Header, Preheader, ExitBlock/*L->getExitBlock()*/,
-                          VMap, M,
-                          F->getSubprogram() != nullptr, Returns, ".ls",
-                          nullptr, nullptr, nullptr);
-
-    assert(Returns.empty() && "Returns cloned when cloning loop.");
-
-    // Use a fast calling convention for the helper.
-    //Helper->setCallingConv(CallingConv::Fast);
-    // Helper->setCallingConv(Header->getParent()->getCallingConv());
-  }
-
-  BasicBlock *NewPreheader = cast<BasicBlock>(VMap[Preheader]);
-  PHINode *NewCanonicalIV = cast<PHINode>(VMap[CanonicalIV]);
-
-  // Rewrite the cloned IV's to start at the start iteration argument.
-  {
-    // Rewrite clone of canonical IV to start at the start iteration
-    // argument.
-    Argument *NewCanonicalIVStart = cast<Argument>(VMap[InputMap[CanonicalIV]]);
-    {
-      int NewPreheaderIdx = NewCanonicalIV->getBasicBlockIndex(NewPreheader);
-      assert(isa<Constant>(NewCanonicalIV->getIncomingValue(NewPreheaderIdx)) &&
-             "Cloned canonical IV does not inherit a constant value from cloned preheader.");
-      NewCanonicalIV->setIncomingValue(NewPreheaderIdx, NewCanonicalIVStart);
-    }
-
-    // Rewrite other cloned IV's to start at their value at the start
-    // iteration.
-    const SCEV *StartIterSCEV = SE.getSCEV(NewCanonicalIVStart);
-    DEBUG(dbgs() << "StartIterSCEV: " << *StartIterSCEV << "\n");
-    for (PHINode *IV : IVs) {
-      if (CanonicalIV == IV) continue;
-
-      // Get the value of the IV at the start iteration.
-      DEBUG(dbgs() << "IV " << *IV);
-      const SCEV *IVSCEV = SE.getSCEV(IV);
-      DEBUG(dbgs() << " (SCEV " << *IVSCEV << ")");
-      const SCEVAddRecExpr *IVSCEVAddRec = cast<const SCEVAddRecExpr>(IVSCEV);
-      const SCEV *IVAtIter = IVSCEVAddRec->evaluateAtIteration(StartIterSCEV, SE);
-      DEBUG(dbgs() << " expands at iter " << *StartIterSCEV <<
-            " to " << *IVAtIter << "\n");
-
-      // NOTE: Expanded code should not refer to other IV's.
-      Value *IVStart = Exp.expandCodeFor(IVAtIter, IVAtIter->getType(),
-                                         NewPreheader->getTerminator());
-
-
-      // Set the value that the cloned IV inherits from the cloned preheader.
-      PHINode *NewIV = cast<PHINode>(VMap[IV]);
-      int NewPreheaderIdx = NewIV->getBasicBlockIndex(NewPreheader);
-      assert(isa<Constant>(NewIV->getIncomingValue(NewPreheaderIdx)) &&
-             "Cloned IV does not inherit a constant value from cloned preheader.");
-      NewIV->setIncomingValue(NewPreheaderIdx, IVStart);
-    }
-
-    // Remap the newly added instructions in the new preheader to use
-    // values local to the helper.
-    for (Instruction &II : *NewPreheader)
-      RemapInstruction(&II, VMap, RF_IgnoreMissingLocals,
-                       /*TypeMapper=*/nullptr, /*Materializer=*/nullptr);
-  }
-
-  // If the loop limit is constant, then rewrite the loop latch
-  // condition to use the end-iteration argument.
-  if (isa<Constant>(LimitVar)) {
-    CmpInst *HelperCond = cast<CmpInst>(VMap[NewCond]);
-    assert(HelperCond->getOperand(1) == LimitVar);
-    IRBuilder<> Builder(HelperCond);
-    Value *NewHelperCond = Builder.CreateICmpULT(HelperCond->getOperand(0),
-                                                 VMap[InputMap[LimitVar]]);
-    HelperCond->replaceAllUsesWith(NewHelperCond);
-    HelperCond->eraseFromParent();
-  }
-
-  // For debugging:
-  BasicBlock *NewHeader = cast<BasicBlock>(VMap[Header]);
-  SerializeDetachedCFG(cast<DetachInst>(NewHeader->getTerminator()), nullptr);
-  {
-    Value* v = &*Helper->arg_begin();
-    auto UI = v->use_begin(), E = v->use_end();
-    for (; UI != E;) {
-      Use &U = *UI;
-      ++UI;
-      auto *Usr = dyn_cast<Instruction>(U.getUser());
-      Usr->moveBefore(Helper->getEntryBlock().getTerminator());
-
-      auto UI2 = Usr->use_begin(), E2 = Usr->use_end();
-      for (; UI2 != E2;) {
-        Use &U2 = *UI2;
-        ++UI2;
-        auto *Usr2 = dyn_cast<Instruction>(U2.getUser());
-        Usr2->moveBefore(Helper->getEntryBlock().getTerminator());
-      }
-    }
-  }
-
-  if (verifyFunction(*Helper, &dbgs()))
-    return false;
-
-  // Add call to new helper function in original function.
-  {
-    // Setup arguments for call.
-    SetVector<Value*> TopCallArgs;
-    // Add start iteration 0.
-    assert(CanonicalSCEV->getStart()->isZero() &&
-           "Canonical IV does not start at zero.");
-    TopCallArgs.insert(ConstantInt::get(CanonicalIV->getType(), 0));
-    // Add loop limit.
-    TopCallArgs.insert(LimitVar);
-    // Add grainsize.
-    //TopCallArgs.insert(GrainVar);
-    // Add the rest of the arguments.
-    for (Value *V : BodyInputs)
-      TopCallArgs.insert(V);
-
-    // Create call instruction.
-    IRBuilder<> Builder(Preheader->getTerminator());
-
-    llvm::Function* F;
-    if( ((llvm::IntegerType*)LimitVar->getType())->getBitWidth() == 32 )
-      F = CILKRTS_FUNC(cilk_for_32, *M);
-    else {
-      assert( ((llvm::IntegerType*)LimitVar->getType())->getBitWidth() == 64 );
-      F = CILKRTS_FUNC(cilk_for_64, *M);
-    }
-    llvm::Value* args[] = {
-      Builder.CreatePointerCast(Helper, F->getFunctionType()->getParamType(0)),
-      Builder.CreatePointerCast(closure, F->getFunctionType()->getParamType(1)),
-      LimitVar,
-      ConstantInt::get(IntegerType::get(F->getContext(), sizeof(int)*8),0)
-    };
-
-    /*CallInst *TopCall = */Builder.CreateCall(F, args);
-
-    // Use a fast calling convention for the helper.
-    //TopCall->setCallingConv(CallingConv::Fast);
-    // TopCall->setCallingConv(Helper->getCallingConv());
-    //TopCall->setDebugLoc(Header->getTerminator()->getDebugLoc());
-    // // Update CG graph with the call we just added.
-    // CG[F]->addCalledFunction(TopCall, CG[Helper]);
-  }
-
-  ++LoopsConvertedToCilkABI;
 
   unlinkLoop();
